backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
apispec==6.8.4
attrs==25.4.0
blinker==1.9.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
cryptography==46.0.2
distro==1.9.0
fastuuid==0.13.5
filelock==3.20.0
Flask==3.1.2
flask-smorest==0.46.2
Flask-SQLAlchemy==3.1.1
frozenlist==1.8.0
fsspec==2025.9.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
litellm==1.78.0
MarkupSafe==3.0.3
marshmallow==4.0.1
multidict==6.7.0
openai==2.3.0
packaging==25.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
propcache==0.4.1
pycparser==2.23
pydantic==2.12.0
pydantic_core==2.41.1
pypdfium2==4.30.0
python-dotenv==1.1.1
PyYAML==6.0.3
referencing==0.36.2
regex==2025.9.18
reportlab==4.4.4
requests==2.32.5
rpds-py==0.27.1
sniffio==1.3.1
SQLAlchemy==2.0.44
tiktoken==0.12.0
tokenizers==0.22.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
watchdog==6.0.0
webargs==8.7.0
Werkzeug==3.1.3
yarl==1.22.0
zipp==3.23.0
